+++ src/renderer/components/argo-config/argo-config-dialog.module.d.scss.ts
@@ -4,5 +4,8 @@ declare const classNames: {
   readonly dialogActions: "dialogActions";
   readonly dialogError: "dialogError";
   readonly textArea: "textArea";
+  readonly kvRow: "kvRow";
+  readonly kvKey: "kvKey";
+  readonly kvValue: "kvValue";
 };
 export = classNames;