+++ src/renderer/components/argo-config/argo-config-dialog.module.scss.d.ts
@@ -1,8 +1,0 @@
-declare const classNames: {
-  readonly dialogContent: "dialogContent";
-  readonly dialogHeader: "dialogHeader";
-  readonly dialogActions: "dialogActions";
-  readonly dialogError: "dialogError";
-  readonly textArea: "textArea";
-};
-export = classNames;