+++ src/renderer/components/error-page.module.scss.d.ts
@@ -1,5 +1,0 @@
-declare const classNames: {
-  readonly errorPage: "errorPage";
-  readonly errorMessage: "errorMessage";
-};
-export = classNames;