+++ src/renderer/pages/argo-landing-page.module.scss.d.ts
@@ -1,8 +1,0 @@
-declare const classNames: {
-  readonly root: "root";
-  readonly title: "title";
-  readonly lead: "lead";
-  readonly list: "list";
-  readonly hint: "hint";
-};
-export = classNames;